+++ DarmuhsTerminalCommands/CommandHandling/CostCommands.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using static OpenLib.CoreMethods.AddingThings;
 using UnityEngine;
+using OpenLib.Common;
 
 namespace TerminalStuff
 {
@@ -19,7 +20,6 @@ namespace TerminalStuff
         internal static string currentPackName;
         internal static string buyPackName;
         internal static Dictionary<Item, int> itemsIndexed = [];
-        internal static TerminalNode dummyNode = CreateDummyNode("costcommands_dummy", true, "");
 
         internal static bool CheckUpgradeStatus(ref bool upgrade, string itemName)
         {
@@ -344,8 +344,7 @@ namespace TerminalStuff
             else
             {
                 Plugin.MoreLogs("not enough credits to purchase, sending to cannot afford display");
-                dummyNode.displayText = $"You cannot afford the {currentPackName} PurchasePack ({itemCount} items).\r\n\r\n\tTotal Cost: ■<color=#BD3131>{totalCost}</color>\r\n\r\n";
-                Plugin.instance.Terminal.LoadNewNode(dummyNode);
+                Plugin.instance.escapeConfirmation = true;
                 return $"You cannot afford the {currentPackName} PurchasePack ({itemCount} items).\r\n\r\n\tTotal Cost: ■<color=#BD3131>{totalCost}</color>\r\n\r\n";
             }
         }